:truck: [#3489] Move BAG client into Kadaster contrib package
