+++ src/openforms/contrib/kadaster/clients/__init__.py
@@ -1,7 +1,13 @@
+from openforms.contrib.bag.models import BAGConfig
 from zgw_consumers_ext.api_client import ServiceClientFactory
 
 from ..models import KadasterApiConfig
+from .bag import BAGClient
 from .locatieserver import LocatieServerClient
+
+
+class NoServiceConfigured(RuntimeError):
+    pass
 
 
 def get_locatieserver_client() -> LocatieServerClient:
@@ -11,3 +17,12 @@ def get_locatieserver_client() -> LocatieServerClient:
     assert (service := config.search_service)
     service_client_factory = ServiceClientFactory(service)
     return LocatieServerClient.configure_from(service_client_factory)
+
+
+def get_bag_client() -> BAGClient:
+    config = BAGConfig.get_solo()
+    assert isinstance(config, BAGConfig)
+    if not (service := config.bag_service):
+        raise NoServiceConfigured("No BAG service configured!")
+    service_client_factory = ServiceClientFactory(service)
+    return BAGClient.configure_from(service_client_factory)
